budget_forecast_app/requirements.txt
# ===========================
# 🧩 Core Framework
# ===========================
Django==4.2.25
django-vite==3.0.1

# Core data science libraries
pandas>=3.0.1
numpy>=1.21.0

# Machine learning libraries
scikit-learn>=1.1.0

# Interactive visualization
plotly>=5.15.0

# Optional: For Jupyter notebook support
# jupyter>=1.0.0
# ipywidgets>=8.0.0

prophet == 1.3.0

celery>=5.3.0
redis>=5.0.0

psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

djangorestframework
django-cors-headers
djangorestframework-simplejwt

xgboost>=2.0.0
catboost>=1.2.0
